ab_example_contract_ft::ffi::new

Function example_ft_new

#[no_mangle]
pub unsafe extern "C" fn example_ft_new(
    args: NonNull<InternalArgs<'_>>,
) -> ExitCode

FFI interface into a method, called by the host.

NOTE: Calling this function directly shouldn’t be necessary except maybe in contract’s own tests.

Safety

Caller must ensure the provided pointer corresponds to expected ABI.